news/requirements.txt
khl.py
pytz
pydantic
openai
python-dotenv
schedule
requests
datetime
asyncio